refactored some folders. finally got getStaticProps to work.
