+++ src/component/display/display.tsx
@@ -1,7 +1,9 @@
 import styles from "./display.module.css";
 
 //const Display = () => {
-const Display = ({ projects }: any) => {
+const Display = ({ projects = [] }: any) => {
+
+  console.log("projects loaded", projects);
   return (
       <div className={styles['column']}>
         {
